[refactor] (stellar) move common HD test to helper.ts
